Fixed warning: conversion from float to int
git-svn-id: http://hcm-lab.de/public/Horde3D/trunk@81 5ce291ac-9df0-446f-9e4f-d57731c4dda7

+++ Tools/GameEngine/src/GameAnimationComponent/MoveAnimComponent.cpp
@@ -238,7 +238,7 @@ void MoveAnimComponent::update(float fps)
 		{
 			if(!m_randSeed)
 			{
-				srand( fps * 100000 );
+				srand( (unsigned int) (fps * 100000) );
 				m_randSeed = true;
 			}
 			int i = rand() % m_idleAnimCount;
